rename map asset utils
