+++ src/map-assets-utils.js
